simplify structure; setting default axes
