+++ pyalluv/__init__.py
@@ -1,6 +1,6 @@
 from .plotting import Alluvial
-from .clusters import Cluster
-from .fluxes import Flux
+from .plotting import _Cluster as Cluster
+from .plotting import _Flux as Flux
 
 # backwards compatibility for obsolete version
 SankeyPlot = Alluvial
