+++ pyalluv/plotting.py
@@ -2,8 +2,579 @@ from __future__ import division, absolute_import, unicode_literals
 from matplotlib.collections import PatchCollection
 import matplotlib.pyplot as plt
 import matplotlib.dates as mdates
+from matplotlib.path import Path
+import matplotlib.patches as patches
 from datetime import datetime
 from bisect import bisect_left
+
+
+class _Cluster(object):
+    r"""
+    This class defines the cluster objects for an alluvial diagram.
+
+    Note
+    -----
+
+    The vertical position of a cluster will be set when creating a
+    :class:`~pyalluv.plotting.Alluvial`.
+
+    Parameters
+    -----------
+
+    height: float, int
+      The cluster size which will translate into the height of this cluster.
+    anchor: float (default=None)
+      Set the anchor position. Either only the horizontal position, both
+      :math:`(x, y)` or nothing can be provided.
+    width: float (default=1.0)
+      Set the cluster width.
+    label: str (default=None)
+      The label for this cluster, that can be shown in the diagram
+    \**kwargs optional parameter:
+      x_anchor: ``'center'``, ``'left'`` or ``'right'`` (default='center')
+        Determine where the anchor position is relative to the rectangle
+        that will represent this cluster. Options are either the left or
+        right corner or centered:
+      linewidth: float (default=0.0)
+        Set the width of the line surrounding a cluster.
+      label_margin: tuple(horizontal, vertical)
+        Sets horizontal and vertical margins for the label of a cluster.
+
+    Attributes
+    -----------
+    x_pos: float
+      Horizontal position of the cluster anchor.
+    y_pos: float
+      Vertical position of the cluster center.
+    x_anchor: str
+      Anchor position relative to the rectangle representing the cluster.
+      Possible values are: ``'center'``, ``'left'`` or ``'right'``.
+    height: float
+      Size of the cluster that will determine its height in the diagram.
+    width: float
+      Width of the cluster. In the same units as ``x_pos``.
+    label: str
+      Label, id or name of the cluster.
+    in_fluxes: list[:class:`~pyalluv.fluxes.Flux`]
+      All incoming fluxes of this cluster.
+    out_fluxes: list[:class:`~pyalluv.fluxes.Flux`]
+      All outgoing fluxes of this cluster.
+
+    """
+    def __init__(self, height, anchor=None, width=1.0, label=None, **kwargs):
+        self._interp_steps = kwargs.pop('_interpolation_steps', 1)
+        self.x_anchor = kwargs.pop('x_anchor', 'center')
+        self.label = label
+        self.label_margin = kwargs.pop('label_margin', None)
+        self._closed = kwargs.pop('closed', False)
+        self._readonly = kwargs.pop('readonly', False)
+        self.patch_kwargs = kwargs
+        self.patch_kwargs['lw'] = self.patch_kwargs.pop(
+            'linewidth', self.patch_kwargs.pop('lw', 0.0)
+        )
+        if isinstance(height, (list, tuple)):
+            self.height = len(height)
+        else:
+            self.height = height
+        self.width = width
+        if isinstance(anchor, (list, tuple)):
+            x_coord, y_coord = anchor
+        else:
+            x_coord, y_coord = anchor, None
+        self = self.set_x_pos(x_coord).set_y_pos(y_coord)
+
+        # init the in and out fluxes:
+        self.out_fluxes = []
+        self.in_fluxes = []
+        self.in_margin = {
+            'bottom': 0,
+            'top': 0
+        }
+        self.out_margin = {
+            'bottom': 0,
+            'top': 0
+        }
+        # ref points to add fluxes
+        self.in_ = None
+        self.out_ = None
+
+    def set_x_pos(self, x_pos):
+        r"""
+        Set the horizontal position of a cluster.
+
+        The position is set according to the value provided in ``x_pos`` and
+        ``self.x_anchor``.
+
+        Parameters
+        -----------
+        x_pos: float
+          Horizontal position of the anchor for the cluster.
+
+        Returns
+        --------
+        self: :class:`.Cluster`
+          with new property ``x_pos``.
+
+        """
+        self.x_pos = x_pos
+        if self.x_pos is not None:
+            self.x_pos -= 0.5 * self.width
+            if self.x_anchor == 'left':
+                self.x_pos += 0.5 * self.width
+            elif self.x_anchor == 'right':
+                self.x_pos -= 0.5 * self.width
+
+        return self
+
+    def get_patch(self, **kwargs):
+        _kwargs = dict(kwargs)
+        _kwargs.update(self.patch_kwargs)
+        self.set_in_out_anchors()
+
+        vertices = [
+            (self.x_pos, self.y_pos),
+            (self.x_pos, self.y_pos + self.height),
+            (self.x_pos + self.width, self.y_pos + self.height),
+            (self.x_pos + self.width, self.y_pos),
+            # this is just ignored as the code is CLOSEPOLY
+            (self.x_pos, self.y_pos)
+        ]
+        codes = [
+            Path.MOVETO,
+            Path.LINETO,
+            Path.LINETO,
+            Path.LINETO,
+            Path.CLOSEPOLY
+        ]
+
+        return patches.PathPatch(
+            Path(
+                vertices,
+                codes,
+                self._interp_steps,
+                self._closed,
+                self._readonly
+            ),
+            **_kwargs
+        )
+
+    def set_loc_out_fluxes(self,):
+        for out_flux in self.out_fluxes:
+            in_loc = None
+            out_loc = None
+            if out_flux.target_cluster is not None:
+                if self.mid_height > out_flux.target_cluster.mid_height:
+                    # draw to top
+                    if self.mid_height >= \
+                            out_flux.target_cluster.in_['top'][1]:
+                        # draw from bottom to in top
+                        out_loc = 'bottom'
+                        in_loc = 'top'
+                    else:
+                        # draw from top to top
+                        out_loc = 'top'
+                        in_loc = 'top'
+                else:
+                    # draw to bottom
+                    if self.mid_height <= \
+                            out_flux.target_cluster.in_['bottom'][1]:
+                        # draw from top to bottom
+                        out_loc = 'top'
+                        in_loc = 'bottom'
+                    else:
+                        # draw form bottom to bottom
+                        out_loc = 'bottom'
+                        in_loc = 'bottom'
+            else:
+                out_flux.out_loc = out_flux.out_flux_vanish
+            out_flux.in_loc = in_loc
+            out_flux.out_loc = out_loc
+
+    def sort_out_fluxes(self,):
+        _top_fluxes = [
+            (i, self.out_fluxes[i])
+            for i in range(len(self.out_fluxes))
+            if self.out_fluxes[i].out_loc == 'top'
+        ]
+        _bottom_fluxes = [
+            (i, self.out_fluxes[i])
+            for i in range(len(self.out_fluxes))
+            if self.out_fluxes[i].out_loc == 'bottom'
+        ]
+        if _top_fluxes:
+            sorted_top_idx, _fluxes_top = zip(*sorted(
+                _top_fluxes,
+                key=lambda x: x[1].target_cluster.mid_height
+                if x[1].target_cluster
+                else -10000,
+                reverse=True
+            ))
+        else:
+            sorted_top_idx = []
+        if _bottom_fluxes:
+            sorted_bottom_idx, _fluxes_bottom = zip(*sorted(
+                _bottom_fluxes,
+                key=lambda x: x[1].target_cluster.mid_height
+                if x[1].target_cluster
+                else -10000,
+                reverse=False
+            ))
+        else:
+            sorted_bottom_idx = []
+        sorted_idx = list(sorted_top_idx) + list(sorted_bottom_idx)
+        self.out_fluxes = [self.out_fluxes[i] for i in sorted_idx]
+
+    def sort_in_fluxes(self,):
+        _top_fluxes = [
+            (i, self.in_fluxes[i])
+            for i in range(len(self.in_fluxes))
+            if self.in_fluxes[i].in_loc == 'top'
+        ]
+        _bottom_fluxes = [
+            (i, self.in_fluxes[i])
+            for i in range(len(self.in_fluxes))
+            if self.in_fluxes[i].in_loc == 'bottom'
+        ]
+        if _top_fluxes:
+            sorted_top_idx, _fluxes_top = zip(*sorted(
+                _top_fluxes,
+                key=lambda x: x[1].source_cluster.mid_height
+                if x[1].source_cluster
+                else -10000,
+                reverse=True
+            ))
+        else:
+            sorted_top_idx = []
+        if _bottom_fluxes:
+            sorted_bottom_idx, _fluxes_bottom = zip(*sorted(
+                _bottom_fluxes,
+                key=lambda x: x[1].source_cluster.mid_height
+                if x[1].source_cluster
+                else -10000,
+                reverse=False
+            ))
+        else:
+            sorted_bottom_idx = []
+        sorted_idx = list(sorted_top_idx) + list(sorted_bottom_idx)
+        self.in_fluxes = [self.in_fluxes[i] for i in sorted_idx]
+
+    def get_loc_out_flux(self, flux_width, out_loc, in_loc):
+        anchor_out = (
+            self.out_[out_loc][0],
+            self.out_[out_loc][1] + self.out_margin[out_loc] + (flux_width if in_loc == 'bottom' else 0)
+        )
+        top_out = (
+            self.out_[out_loc][0],
+            self.out_[out_loc][1] + self.out_margin[out_loc] + (flux_width if in_loc == 'top' else 0)
+        )
+        self.out_margin[out_loc] += flux_width
+        return anchor_out, top_out
+
+    def set_anchor_out_fluxes(self,):
+        for out_flux in self.out_fluxes:
+            out_width = out_flux.flux_width \
+                if out_flux.out_loc == 'bottom' else - out_flux.flux_width
+            out_flux.anchor_out, out_flux.top_out = self.get_loc_out_flux(
+                out_width, out_flux.out_loc, out_flux.in_loc
+            )
+
+    def set_anchor_in_fluxes(self,):
+        for in_flux in self.in_fluxes:
+            in_width = in_flux.flux_width \
+                if in_flux.in_loc == 'bottom' else - in_flux.flux_width
+            in_flux.anchor_in, in_flux.top_in = self.get_loc_in_flux(
+                in_width, in_flux.out_loc, in_flux.in_loc
+            )
+
+    def get_loc_in_flux(self, flux_width, out_loc, in_loc):
+        anchor_in = (
+            self.in_[in_loc][0],
+            self.in_[in_loc][1] + self.in_margin[in_loc] + (flux_width if out_loc == 'bottom' else 0)
+        )
+        top_in = (
+            self.in_[in_loc][0],
+            self.in_[in_loc][1] + self.in_margin[in_loc] + (flux_width if out_loc == 'top' else 0)
+        )
+        self.in_margin[in_loc] += flux_width
+        return anchor_in, top_in
+
+    def set_mid_height(self, mid_height):
+        self.mid_height = mid_height
+        if self.mid_height is not None:
+            self.y_pos = self.mid_height - 0.5 * self.height
+            self.set_in_out_anchors()
+        else:
+            self.y_pos = None
+
+    def set_y_pos(self, y_pos):
+        self.y_pos = y_pos
+        if self.y_pos is not None:
+            self.mid_height = self.y_pos + 0.5 * self.height
+            self.set_in_out_anchors()
+        else:
+            self.mid_height = None
+
+        return self
+
+    def set_in_out_anchors(self,):
+        """
+        This sets the proper anchor points for fluxes to enter/leave
+        """
+        # if self.y_pos is None or self.mid_height is None:
+        #     self.set_y_pos()
+
+        self.in_ = {
+            'bottom': (self.x_pos, self.y_pos),  # left, bottom
+            'top': (self.x_pos, self.y_pos + self.height)  # left, top
+        }
+        self.out_ = {
+            # right, top
+            'top': (self.x_pos + self.width, self.y_pos + self.height),
+            'bottom': (self.x_pos + self.width, self.y_pos)  # right,bottom
+        }
+
+
+class _Flux(object):
+    r"""
+
+    Parameters
+    -----------
+    relative_flux: bool
+      If ``True`` the fraction of the height of parameter `source_cluster`
+      is taken, if the source_cluster is none, then the
+      relative height form the target_cluster is taken.
+    source_cluster: :class:`pyalluv.clusters.Cluster` (default=None)
+      Cluster from which the flux originates.
+    target_cluster: :class:`pyalluv.clusters.Cluster` (default=None)
+      Cluster into which the flux leads.
+    \**kwargs optional parameter:
+      interpolation_steps:
+
+      out_flux_vanish: str (default='top')
+
+      default_fc: (default='gray')
+
+      default_ec: (default='gray')
+
+      default_alpha: int (default=0.3)
+
+      closed
+      readonly
+      facecolors
+      edgecolors
+      linewidths
+      linestyles
+      antialiaseds
+
+    Attributes
+    -----------
+
+    flux: float
+      The size of the flux which will translate to the height of the flux in
+      the Alluvial diagram.
+    source_cluster: :class:`pyalluv.clusters.Cluster` (default=None)
+      Cluster from which the flux originates.
+    target_cluster: :class:`pyalluv.clusters.Cluster` (default=None)
+      Cluster into which the flux leads.
+    """
+    def __init__(
+            self, flux,
+            source_cluster=None, target_cluster=None,
+            relative_flux=False,
+            **kwargs):
+        self._interp_steps = kwargs.pop('interpolation_steps', 1)
+        self.out_flux_vanish = kwargs.pop('out_flux_vanish', 'top')
+        self.default_fc = kwargs.pop('default_fc', 'gray')
+        self.default_ec = kwargs.pop('default_ec', 'gray')
+        self.default_alpha = kwargs.pop('default_alpha', 0.3)
+        # self.default_alpha = kwargs.pop(
+        #         'default_alpha',
+        #         kwargs.get('alpha', {}).pop('default', 0.3)
+        #         )
+        self.closed = kwargs.pop('closed', False)
+        self.readonly = kwargs.pop('readonly', False)
+        self.patch_kwargs = kwargs
+        self.patch_kwargs['lw'] = self.patch_kwargs.pop(
+            'linewidth', self.patch_kwargs.pop('lw', 0.0)
+        )
+
+        if isinstance(flux, (list, tuple)):
+            self.flux = len(flux)
+        else:
+            self.flux = flux
+        self.relative_flux = relative_flux
+        self.source_cluster = source_cluster
+        self.target_cluster = target_cluster
+        if self.source_cluster is not None:
+            if self.relative_flux:
+                self.flux_width = self.flux * self.source_cluster.height
+            else:
+                self.flux_width = self.flux
+        else:
+            if self.target_cluster is not None:
+                if self.relative_flux:
+                    self.flux_width = self.flux * self.target_cluster.height
+                else:
+                    self.flux_width = self.flux
+        # append the flux to the clusters
+        if self.source_cluster is not None:
+            self.source_cluster.out_fluxes.append(self)
+        if self.target_cluster is not None:
+            self.target_cluster.in_fluxes.append(self)
+
+    def get_patch(self, **kwargs):
+        _kwargs = dict(kwargs)
+        _to_in_kwargs = {}
+        _to_out_kwargs = {}
+        for kw in _kwargs:
+            if kw.startswith('in_'):
+                _to_in_kwargs[kw[3:]] = _kwargs.pop(kw)
+            elif kw.startswith('out_'):
+                _to_out_kwargs[kw[3:]] = _kwargs.pop(kw)
+        # update with Flux specific styling
+        _kwargs.update(self.patch_kwargs)
+        for _color in ['facecolor', 'edgecolor']:
+            _set_color = _kwargs.pop(_color, None)
+            _set_alpha = _kwargs.pop('alpha', None)
+            if isinstance(_set_alpha, (int, float)):
+                _kwargs['alpha'] = _set_alpha
+                _set_alpha = None
+            color_is_set = False
+            if _set_color == 'source_cluster' or _set_color == 'cluster':
+                from_cluster = self.source_cluster
+                color_is_set = True
+            elif _set_color == 'target_cluster':
+                from_cluster = self.target_cluster
+                color_is_set = True
+            elif isinstance(_set_color, str) and '__' in _set_color:
+                which_cluster, flow_type = _set_color.split('__')
+                if which_cluster == 'target_cluster':
+                    from_cluster = self.target_cluster
+                else:
+                    from_cluster = self.source_cluster
+                if flow_type == 'migration' \
+                        and self.source_cluster.patch_kwargs.get(_color) \
+                        != self.target_cluster.patch_kwargs.get(_color):
+                    color_is_set = True
+                    if _set_alpha:
+                        _kwargs['alpha'] = _set_alpha.get(
+                            'migration', _set_alpha.get('default', self.default_alpha)
+                        )
+                elif flow_type == 'reside'  \
+                        and self.source_cluster.patch_kwargs.get(_color) \
+                        == self.target_cluster.patch_kwargs.get(_color):
+                    color_is_set = True
+                    if _set_alpha:
+                        _kwargs['alpha'] = _set_alpha.get(
+                            'reside', _set_alpha.get('default', self.default_alpha)
+                        )
+                else:
+                    _set_color = None
+            if color_is_set:
+                _kwargs[_color] = from_cluster.patch_kwargs.get(
+                    _color, None
+                )
+
+            # set it back
+            else:
+                _kwargs[_color] = _set_color
+                if _set_color is None:
+                    if _color == 'facecolor':
+                        _kwargs[_color] = self.default_fc
+                    elif _color == 'edgecolor':
+                        _kwargs[_color] = self.default_ec
+                if _set_alpha:
+                    _kwargs['alpha'] = _set_alpha.get('default', self.default_alpha)
+        # line below is probably not needed as alpha is set with the color
+        _kwargs['alpha'] = _kwargs.get('alpha', self.default_alpha)
+        # set in/out only flux styling
+        _in_kwargs = dict(_kwargs)
+        _in_kwargs.update(_to_in_kwargs)
+        _out_kwargs = dict(_kwargs)
+        _out_kwargs.update(_to_out_kwargs)
+
+        _dist = None
+        if self.out_loc is not None:
+            if self.in_loc is not None:
+                _dist = 2 / 3 * (
+                    self.target_cluster.in_['bottom'][0] - self.source_cluster.out_['bottom'][0]
+                )
+            else:
+                _dist = 2 * self.source_cluster.width
+                _kwargs = _out_kwargs
+        else:
+            if self.in_loc is not None:
+                _kwargs = _in_kwargs
+            else:
+                raise Exception('Flux with neither source nor target cluster')
+
+        # now complete the path points
+        if self.anchor_out is not None:
+            anchor_out_inner = (
+                self.anchor_out[0] - 0.5 * self.source_cluster.width,
+                self.anchor_out[1]
+            )
+            dir_out_anchor = (self.anchor_out[0] + _dist, self.anchor_out[1])
+        else:
+            # TODO set to form vanishing flux
+            # anchor_out = anchor_out_inner =
+            # dir_out_anchor =
+            pass
+        if self.top_out is not None:
+            top_out_inner = (
+                self.top_out[0] - 0.5 * self.source_cluster.width,
+                self.top_out[1]
+            )
+            # 2nd point 2/3 of distance between clusters
+            dir_out_top = (self.top_out[0] + _dist, self.top_out[1])
+        else:
+            # TODO set to form vanishing flux
+            # top_out = top_out_inner =
+            # dir_out_top =
+            pass
+        if self.anchor_in is not None:
+            anchor_in_inner = (
+                self.anchor_in[0] + 0.5 * self.target_cluster.width,
+                self.anchor_in[1]
+            )
+            dir_in_anchor = (self.anchor_in[0] - _dist, self.anchor_in[1])
+        else:
+            # TODO set to form new in flux
+            # anchor_in = anchor_in_inner =
+            # dir_in_anchor =
+            pass
+        if self.top_in is not None:
+            top_in_inner = (
+                self.top_in[0] + 0.5 * self.target_cluster.width,
+                self.top_in[1]
+            )
+            dir_in_top = (self.top_in[0] - _dist, self.top_in[1])
+        else:
+            # TODO set to form new in flux
+            # top_in = top_in_inner =
+            # dir_in_top =
+            pass
+
+        vertices = [
+            self.anchor_out,
+            dir_out_anchor, dir_in_anchor, self.anchor_in,
+            anchor_in_inner, top_in_inner, self.top_in,
+            dir_in_top, dir_out_top, self.top_out,
+            top_out_inner, anchor_out_inner,
+            self.anchor_out
+        ]
+        codes = [
+            Path.MOVETO,
+            Path.CURVE4, Path.CURVE4, Path.CURVE4,
+            Path.LINETO, Path.LINETO, Path.LINETO,
+            Path.CURVE4, Path.CURVE4, Path.CURVE4,
+            Path.LINETO, Path.LINETO,
+            Path.CLOSEPOLY
+        ]
+        _path = Path(vertices, codes, self._interp_steps, self.closed, self.readonly)
+
+        flux_patch = patches.PathPatch(_path, **_kwargs)
+        return flux_patch
 
 
 class Alluvial(object):
@@ -16,7 +587,7 @@ class Alluvial(object):
         `Wikipedia (23/1/2021) <https://en.wikipedia.org/wiki/Alluvial_diagram>`_
     """
     def __init__(
-        self, clusters, axes, y_pos='overwrite', cluster_w_spacing=1,
+        self, clusters, ax=None, y_pos='overwrite', cluster_w_spacing=1,
         cluster_kwargs={}, flux_kwargs={}, label_kwargs={},
         **kwargs
     ):
@@ -52,8 +623,9 @@ class Alluvial(object):
             :obj:`.Cluster` (*value*) for a horizontal position (*key*), e.g.
             ``{1.0: [c11, c12, ...], 2.0: [c21, c22, ...], ...}``.
 
-        axes: :class:`matplotlib.axes.Axes`
-          Axes to draw an Alluvial diagram on.
+        ax: `~.axes.Axes`
+          Axes onto which the Alluvial diagram should be drawn.
+          If *ax* is not provided a new Axes instance will be created.
         y_pos: str
           **options:** ``'overwrite'``, ``'keep'``, ``'complement'``, ``'sorted'``
 
@@ -162,6 +734,13 @@ class Alluvial(object):
         clusters: dict
           Holds for each vertical position a list of :obj:`.Cluster` objects.
         """
+        # create axes if not provided
+        if ax is None:
+            import matplotlib.pyplot as plt
+            fig = plt.figure()
+            # TODO: not sure if specifying the ticks is necessary
+            ax = fig.add_subplot(1, 1, 1, xticks=[], yticks=[])
+
         # if clusters are given in a list of lists (each list is a x position)
         self._set_x_pos = kwargs.get('set_x_pos', True)
         self._redistribute_vertically = kwargs.get(
@@ -201,8 +780,8 @@ class Alluvial(object):
             if self.format_xaxis:
                 locator = mdates.AutoDateLocator(minticks=3)
                 formatter = mdates.AutoDateFormatter(locator)
-                axes.xaxis.set_major_locator(locator)
-                axes.xaxis.set_major_formatter(formatter)
+                ax.xaxis.set_major_locator(locator)
+                ax.xaxis.set_major_formatter(formatter)
             self._x_dates = True
             if (self.x_positions[-1] - self.x_positions[0]).days < 2 * 30:
                 _minor_tick = 'weeks'
@@ -287,20 +866,20 @@ class Alluvial(object):
             cluster_kwargs=self._cluster_kwargs,
             flux_kwargs=self._flux_kwargs
         )
-        axes.add_collection(patch_collection)
+        ax.add_collection(patch_collection)
         if self.with_cluster_labels:
             label_collection = self.get_labelcollection(**label_kwargs)
             if label_collection:
                 for label in label_collection:
-                    axes.annotate(**label)
-        axes.set_xlim(
+                    ax.annotate(**label)
+        ax.set_xlim(
             *self.x_lim
         )
-        axes.set_ylim(
+        ax.set_ylim(
             *self.y_lim
         )
         if self._fill_figure:
-            axes.set_position(
+            ax.set_position(
                 [
                     0.0,
                     self._x_axis_offset,
@@ -309,15 +888,15 @@ class Alluvial(object):
                 ]
             )
         if self._invisible_y:
-            axes.get_yaxis().set_visible(False)
+            ax.get_yaxis().set_visible(False)
         if self._invisible_x:
-            axes.get_xaxis().set_visible(False)
-        axes.spines['right'].set_color('none')
-        axes.spines['left'].set_color('none')
-        axes.spines['top'].set_color('none')
-        axes.spines['bottom'].set_color('none')
+            ax.get_xaxis().set_visible(False)
+        ax.spines['right'].set_color('none')
+        ax.spines['left'].set_color('none')
+        ax.spines['top'].set_color('none')
+        ax.spines['bottom'].set_color('none')
         if isinstance(self.x_positions[0], datetime) and self.format_xaxis:
-            self.set_dates_xaxis(axes, _minor_tick)
+            self.set_dates_xaxis(ax, _minor_tick)
 
     def distribute_clusters(self, x_pos):
         r"""
